pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "delius"
version = "0.0.1"
dependencies = [
  "scikit-learn==1.6.1",
  "matplotlib==3.10.1",
  "tqdm==4.67.1",
  "torch==2.6.0",
  "torchaudio==2.6.0",
  "torchvision==0.21.0",
  "umap-learn==0.5.7"
]
description = "Deep clustering for the visual arts."
authors = [
  { name = "Gianmarco Turchiano" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
requires-python = "~=3.10"

[project.optional-dependencies]
dev = [
  "dvc==3.59.1",
  "dvc-s3==3.2.0",
  "mlflow==2.20.3",
  "python-dotenv==1.0.1"
]

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.ruff.lint.isort]
known_first_party = ["delius"]
force_sort_within_sections = true
